+++ apis/neptune/v1alpha2/zz_clustersnapshot_types.go
@@ -26,7 +26,7 @@ import (
 )
 
 type ClusterSnapshotObservation struct {
-	AllocatedStorage *int64 `json:"allocatedStorage,omitempty" tf:"allocated_storage,omitempty"`
+	AllocatedStorage *float64 `json:"allocatedStorage,omitempty" tf:"allocated_storage,omitempty"`
 
 	AvailabilityZones []*string `json:"availabilityZones,omitempty" tf:"availability_zones,omitempty"`
 
@@ -42,7 +42,7 @@ type ClusterSnapshotObservation struct {
 
 	LicenseModel *string `json:"licenseModel,omitempty" tf:"license_model,omitempty"`
 
-	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	SnapshotType *string `json:"snapshotType,omitempty" tf:"snapshot_type,omitempty"`
